# Speed up interval rounding (backport of #63245)